refactor: moved the builder of the param gen onto it's unique file + integrated the param_server to use it instead
