+++ machine_learning/src/initialization/mod.rs
@@ -1,9 +1,11 @@
+mod builder;
 mod chained;
 mod constant;
 mod error;
 mod param_gen;
 mod random;
 
+pub use builder::ParamGenBuilder;
 pub use chained::ChainedParamGen;
 pub use constant::ConstParamGen;
 pub use error::Result;
